### 2022-07-05 #4546 @editorialbot
[PRE REVIEW]: netrankr: An R package for total, partial, and probabilistic rankings in networks
**Submitting author:** @schochastics (<a href="http://orcid.org/0000-0003-2952-4812">David Schoch</a>)
**Repository:** https://github.com/schochastics/netrankr
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.1
