### 2022-07-05 #4546 @jbytecode
@danielskatz of course, gladly!

@schochastics - Thank you for the submission. The manuscript should include a statement of need section. In this section, you can discuss why we need such a software and what are the alternatives, or opponents, and why people decide to use the software rather than the alternatives, etc. Please take a look at the https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain

Since it is not the rigid requirement, it would be good to show how to install the software, how to use it - with code snippets.

Addition to these, do you have any suggestions for potential reviewers? You can use [the list of people](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822). Please list the usernames without using the @ (at) character, just because we don't want to bother them.

Thank you in advance.
