+++ paper.bib
@@ -15,7 +15,8 @@
   journal = {Journal of Statistical Software},
   volume = {24},
   number = {6},
-  issn = {1548-7660}
+  issn = {1548-7660},
+  doi = {10.18637/jss.v024.i06}
 }
 
 @article{jsaayga-ccrwarpca-15,
@@ -124,7 +125,8 @@
   journal = {Environmental and Ecological Statistics},
   volume = {11},
   number = {2},
-  pages = {199--228}
+  pages = {199--228},
+  doi = {10.1023/b:eest.0000027209.93218.d9}
 }
 
 @article{jmbo-lcpn-01,
