### 2022-07-06 #4546 @schochastics
@jbytecode I have fixed all missing dois and added a statement of need which also discusses alternative R packages.
I would suggest the following reviewers:
MatthewSmith430,
akbaritabar
