### 2022-07-06 #4546 @akbaritabar
> 👋👋👋 Dear @akbaritabar 👋👋👋
>
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**netrankr: An R package for total, partial, and probabilistic rankings in networks**". You can find more information at the top of this Github issue (#4546).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.

I would be delighted to serve as a reviewer.
I must confess that reading the manuscript, my role will be mainly looking at it from an applied social scientist perspective, i.e., ease of use of the package, simplicity and readability of the paper and examples aiding the less specialized reader to understand, be convinced of the use-cases and adopt the package in their toolbox of Network Analysis.

Will get back to you with my review note!
