### 2022-07-06 #4546 @akbaritabar
> Dear @akbaritabar
>
> This is the pre-review issue. In the pre-review issues we always do initial stuff and find suitable reviewers. After starting the review, a new thread will be created. In this review thread, a checklist with 20 reviewing items is created for each reviewer. The process is interactive, that is, the author(s), the reviewers, and the editor always interact, reviewers are welcome to open new issues in the target repo, send pull requests, and any contributions are welcome.

Thanks for the description. I have done a review for JOSS previously and know the process. Thanks.
