### 2022-07-08 #4546 @jbytecode
👋👋👋 Dear @jankatins 👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**netrankr: An R package for total, partial, and probabilistic rankings in networks**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4546).
