### 2022-07-11 #4546 @briatte
> Dear @briatte 👋👋👋
>
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?

Yes, no problem.
