### 2022-08-09 #4563 @jbytecode
> ## Letter to authors (as comment on review issue of JOSS)
> Thanks a lot for the opportunity to review this work and package. The package is nicely documented and many use cases and multiple vignettes and blog posts (http://blog.schochastics.net/post/network-centrality-in-r-introduction/ , http://blog.schochastics.net/post/network-centrality-in-r-neighborhood-inclusion/ , http://blog.schochastics.net/post/network-centrality-in-r-new-ways-of-measuring-centrality/) are provided to users and showcase examples, discuss concepts and why they are implemented and also provide references to literature from author or others.
>
> I must add that my reading of these material and use of the package is mainly "an empirical social scientist's point of view" who uses the proposed techniques but could not contribute to their technical development. Hence, I cannot evaluate the technical aspects of the package and its performance very well and leave that to editor and other reviewers.
>
> The JOSS paper could benefit from including a little bit more detail and examples which are already provided in the package's website and would complement the short paper. It will help the paper be more independent.
>
> Please see below under checklist points only minor suggestions for edits and revisions.
>
> ## Review checklist for @akbaritabar
> ### Conflict of interest
> * [x]  I confirm that I have read the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> ### Code of Conduct
> * [x]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [x]  **Repository:** Is the source code for this software available at the https://github.com/schochastics/netrankr?
> * [ ]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
>
>   * On the package website (http://netrankr.schochastics.net/) MIT license is mentioned and a file is provided stating year and copyright holder's name (as suggested by: https://r-pkgs.org/license.html#key-files) but in the repository on GitHub, only that file is shown instead of MIT license that GitHub usually shows and clicking shows permissions (example of what I intended: https://github.com/openjournals/joss-reviews/blob/master/LICENSE)
> * [x]  **Contribution and authorship:** Has the submitting author (@schochastics) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
> * [x]  **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
>
> ### Functionality
> * [x]  **Installation:** Does installation proceed as outlined in the documentation?
> * [x]  **Functionality:** Have the functional claims of the software been confirmed?
> * [x]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
>   * [x]  Guidelines and rules of thumb are provided on the performance and how it drops for larger networks which is very valuable (http://netrankr.schochastics.net/articles/benchmarks.html)
>
> ### Documentation
> * [x]  **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [x]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>
>   * [x]  In the package website, there are multiple vignettes and tutorials using different examples.
>   * [ ]  It would be great if one of those examples could be included in the paper to make it "stand alone" for those who would read the paper first without trying out the package and vignettes. (a good candidate could be: http://netrankr.schochastics.net/#simple-example)
> * [x]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> * [x]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?
> * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>   [x] There are three issues already created in the repository, and the website (http://netrankr.schochastics.net/index.html) has a button "report a bug" but I couldn't see something on contribution.
>
> ### Software paper
> * [ ]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
>
>   * [x]  Summary presents a very nice definition of "centrality" in networks using examples from different fields.
>   * [ ]  But, a simple and accessible definition of "partial" and "probabilistic ranking" could be added (vignettes already have materials that could be used in the paper to to make it "stand alone" for those who would read the paper first without trying out the package and vignettes). I would suggest using materials already present in package's website (i.e., http://netrankr.schochastics.net/#overview ) or for instance, some of the one-liner questions used in the statement of need would be helpful, e.g., "how likely is a node more central than another one?"
> * [x]  **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
> * [x]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
> * [x]  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
>
>   * Minor point: I would suggest removing citations in the abstract, if possible.
>   * Minor point: "small packages" --> "small package"
> * [x]  **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?
>
> ## A suggestion for JOSS
> * Perhaps it would be nice to use the editorial bot currently used by JOSS and set a reminder for when the review is due, after a reviewer accepts the invitation.

@schochastics - Here are some review notes from one of our reviewers.

@akbaritabar - Since you are editing the old message, we are not be aware of what is new in the thread, could you please provide your next comments as a new posts?

thank you in advance.
