### 2022-08-10 8604f9e
added MIT license file and CoC
