+++ README.md
@@ -126,12 +126,16 @@ P
 ```
 
 [Schoch & Brandes (2016)](https://doi.org/10.1017/S0956792516000401)
-showed that *N*(*u*) ⊆ *N*\[*v*\] (i.e. `P[u,v]=1`) implies
-*c*(*u*) ≤ *c*(*v*) for centrality indices *c*, which are defined via
-specific path algebras. These include many of the well-known measures
-like closeness (and variants), betweenness (and variants) as well as
-many walk-based indices (eigenvector and subgraph centrality, total
-communicability,…).
+showed that
+![N(u) \subseteq N\[v\]](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;N%28u%29%20%5Csubseteq%20N%5Bv%5D "N(u) \subseteq N[v]")
+(i.e. `P[u,v]=1`) implies
+![c(u) \leq c(v)](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;c%28u%29%20%5Cleq%20c%28v%29 "c(u) \leq c(v)")
+for centrality indices
+![c](https://latex.codecogs.com/png.image?%5Cdpi%7B110%7D&space;%5Cbg_white&space;c "c"),
+which are defined via specific path algebras. These include many of the
+well-known measures like closeness (and variants), betweenness (and
+variants) as well as many walk-based indices (eigenvector and subgraph
+centrality, total communicability,…).
 
 Neighborhood-inclusion defines a partial ranking on the set of nodes.
 Each ranking that is in accordance with this partial ranking yields a
@@ -288,3 +292,10 @@ package, consult the following literature:
 > Brandes, Ulrik. (2016). Network Positions. *Methodological
 > Innovations*, **9**, 2059799116630650.
 > ([link](https://dx.doi.org/10.1177/2059799116630650))
+
+## Code of Conduct
+
+Please note that the netrankr project is released with a [Contributor
+Code of
+Conduct](https://contributor-covenant.org/version/2/1/CODE_OF_CONDUCT.html).
+By contributing to this project, you agree to abide by its terms.
