+++ paper.md
@@ -5,7 +5,7 @@ tags:
 - network analysis
 - network centrality
 - partial orders
-date: "06 July 2022"
+date: "10 August 2022"
 output: pdf_document
 bibliography: paper.bib
 affiliations:
@@ -43,7 +43,7 @@ However, there also exist dedicated centrality packages, such as
 `centiserve` [@jsaayga-ccrwarpca-15], `CINNA` [@amj-crpdcinna-19], and `influenceR` [@sa-istqsinn-15].
 The biggest in terms of implemented indices is `centiserve` which includes $33$ indices.
 The primary purpose of `CINNA` is to facilitate the choice of indices by visual 
-and statistical tools. `influenceR` is a comparably small packages which implements 
+and statistical tools. `influenceR` is a comparably small package which implements 
 only a few specialized measures.
 
 The `netrankr` package also offers 
@@ -76,9 +76,110 @@ partial ranking on the vertices of a network.
 
 @sb-rcsn-16 showed that if $c:V \to \mathbb{R}$ is a centrality index, then 
 $$N(u)\subseteq N[v] \implies c(u) \leq c(v)$$
-That is, neighborhood inclusion is preserved by centrality indices and the ranking induced by indices
-can be viewed as linear extensions of the partial ranking induced by neighborhood-inclusion. Analyzing this partial
-ranking thus means that all possible centrality rankings can be analyzed at once.
-More technical details can be found in the dedicated literature [@sb-sninc-15; @b-np-16;@svb-ccicurg-17;@b-cpsn-20].
+That is, neighborhood inclusion is preserved by indices and the ranking can be viewed as linear extensions of the partial ranking induced by neighborhood-inclusion. Analyzing this partial ranking thus means that all possible centrality rankings can be analyzed at once. More technical details can be found in the dedicated literature [@sb-sninc-15; @b-np-16;@svb-ccicurg-17;@b-cpsn-20].
+
+# Example usage
+
+This example briefly explains some of the functionality of the package and the 
+difference to an index driven approach. For more detailed applications see 
+the package vignettes. 
+
+We work with a small graph included in the package.
+``` r
+library(igraph)
+library(netrankr)
+
+data("dbces11")
+```
+
+![](figures/dbces.pdf)
+
+Say we are interested in the most central node of the graph and simply
+compute some standard centrality scores with the `igraph` package.
+``` r
+cent_scores <- data.frame(
+   degree = degree(g),
+   betweenness = round(betweenness(g),4),
+   closeness = round(closeness(g),4),
+   eigenvector = round(eigen_centrality(g)$vector,4),
+   subgraph = round(subgraph_centrality(g),4))
+
+# What are the most central nodes for each index?
+apply(cent_scores,2,which.max)
+#>      degree betweenness   closeness eigenvector    subgraph 
+#>          11           8           6           7          10
+```
+
+
+Each index assigns the highest value to a different
+vertex and it is not clear which may be the correct choice. 
+
+A more general assessment using `netrankr` starts by calculating the neighborhood inclusion preorder.
+
+``` r
+P <- neighborhood_inclusion(g)
+P
+#>    1 2 3 4 5 6 7 8 9 10 11
+#> 1  0 0 1 0 1 1 1 0 0  0  1
+#> 2  0 0 0 1 0 0 0 1 0  0  0
+#> 3  0 0 0 0 1 0 0 0 0  0  1
[…]
