### 2022-08-10 #4563 @schochastics
@akbaritabar Thanks for your comments. Below I address all the open points.

## General checks
**License**
I added an MIT license file to github. This is now also displayed in the right panel.

## Documentation

**example usage**
I added and revised the example from the README

**Community guidelines**
I added code of conduct for contributors which outlines the guidelines of contribution

## Software Paper
**Summary**
I tried to add some more explanations in the example for the terms "partial" and "probabilistic ranking". Happy to include even more but I am wary of the word limit

**Quality of writing**
>Minor point: I would suggest removing citations in the abstract, if possible.

I left the citations because I think the summary is different to an abstract. I also checked some published papers which also had citation. If there is something in the guidelines I missed regarding this, then I will remove the references

>Minor point: "small packages" -> "small package"

[fixed]
