### 2022-08-10 #4563 @jbytecode
@schochastics - our bot recognizes commands if the command is located in the first line of the post.
