### 2022-09-02 #4563 @jbytecode
Dear @alexanderfurnas,

Our first reviewer has completed their tasks. Because we failed to reach our 2nd reviewer, I need a second reviewer for this submission.

Would you be willing to review for this submission for JOSS (Journal of Open Source Software)?

Thank you in advance.
