### 2022-09-18 schochastics/netrankr#10 @corybrunson
non-unique LICENSE file
Currently the root directory contains two license files: LICENSE and LICENSE.md. The first is cited in DESCRIPTION, but the second was created to address another reviewer's request. I believe CRAN requires and only recognizes the first name, so it would be helpful and possibly necessary to combine these into a single file, named 'LICENSE' but with the content of the Markdown file.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4563)._
