### 2022-09-18 schochastics/netrankr#11 @corybrunson
interpretable output in indirect relations vignette
The vignettes in this package are, on the whole, exemplary. I appreciate that each addresses a specific topic so that none grow to impractical length.

The vignette "Indirect relations in networks" has an obvious limitation: A reader who is not executing code along the way cannot see or interpret the results. While too many matrices of indirect relations are calculated to all be printed, it would be helpful for one or two (e.g. for a predefined relation and for a custom relation adapted from it) to be printed and compared (element-wise subtraction or division? juxtaposed visualizations?).

This won't hold up my review, but i believe it would make this functionality more readily accessible. : )

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4563)._
