### 2022-09-18 #4563 @corybrunson
@jbytecode i've encountered a small bug in (i think) how JOSS renders PDFs from 'paper.md'. This submission's paper includes some plain double quotes in R code chunks (e.g. `data("dbces11")`) that are changed to smart quotes (`data(”dbces11”)`) in the PDF. This prevents me from executing the code when copying it from the PDF and pasting it to the console. While i can use the Markdown file to reproduce the example, many learners (who learn about these packages from the published papers) are likely to struggle for a bit before realizing the problem. Could the PDF be generated without making this change?
