### 2022-09-18 schochastics/netrankr#11 @schochastics
Thanks for the suggestion! I'll put this on the list for the next release.
