### 2022-09-18 schochastics/netrankr#10 @corybrunson
Hm. It does, thank you. I see plenty of packages with what i believe are standard licenses in LICENSE alone, but you're certainly right that your approach is common.
