### 2022-09-18 schochastics/netrankr#12 @corybrunson
value of returning rank interval midpoints
This is more interpretative a question and not a review condition. `rank_intervals()` returns not only the intervals of each node's centrality rank range but their midpoints, which are NB'd to be different from the nodes' expected ranks. I gather that they are also not, in general, the nodes' _median_ ranks. Are the midpoints theoretically interesting, or are they just a plotting convenience to anchor the intervals?

In case they are not intrinsically interesting, i would expect them to be more often than not misinterpreted or at least accorded undue importance by analysts, and i would suggest reconsidering them. (I apologize if this is explained in one of the papers; i have not read them thoroughly.)

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4563)._
