### 2022-09-18 schochastics/netrankr#13 @corybrunson
benchmarking code
The benchmarking vignette is compelling (and the availability of approximation methods is impressive), but i am unable to check the results for lack of source code. Certainly the vignette should not itself reproduce this analysis, but i believe the code, or code for a smaller, illustrative benchmarking test, should be included—with the repository if not with the package, e.g. in a folder ignored by the R build.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4563)._
