### 2022-09-18 schochastics/netrankr#15 @corybrunson
documentation for index builder
I expect that the centrality index builder is quite useful for anyone not proficient with R. It is natural to use, so i understand why it has [minimal documentation](https://schochastics.github.io/netrankr/reference/index_builder.html), but i believe it would be helpful to include a sentence there about how the gadget works and what packages it depends on.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4563)._
