### 2022-09-18 schochastics/netrankr#16 @corybrunson
community guidelines
The README invokes and links to the Contributor Code of Conduct, but i was not able to find guidance on how to seek support, report issues, or contribute to the package. (GitHub also [indicates](https://github.com/schochastics/netrankr/community) that contributing guidelines are missing.) Typically this is done in a CONTRIBUTING file in the root or in the .github folder. Such guidelines are a requirement for publication in JOSS.

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4563)._
