### 2022-09-18 schochastics/netrankr#17 @corybrunson
goodpractice alerts
I ran `goodpractice::gp()`, which flagged a few practices that might be changed to make code less fragile (though the quality of the code overall seems solid):

* `sapply()` can return different structures, especially when `simplify` is not set to `FALSE`; `vapply()` is more stable.
* `F` and `T` are variables that can be redefined; `FALSE` and `TRUE` are more stable.
* The single-colon operator `:` can produce unexpected output on edge cases; the `seq_*()` functions are more stable.

It also indicated that only 65% of code lines are covered by tests.

While i follow these practices myself, i leave it to the developers to decide for themselves; they will not affect my review. : )

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/4563)._
