### 2022-09-18 #4563 @corybrunson
@schochastics i have concluded my initial review! The package is the first i've seen of this research program since Bauer's positionality paper from several years ago, and it was fascinating to dive into. Thank you so much for making the work so accessible.

While i made several other suggestions that i hope are valuable, my checklist awaits resolutions only to issues [#13](https://github.com/schochastics/netrankr/issues/13) and [#16](https://github.com/schochastics/netrankr/issues/16).
