### 2022-09-20 5857a13
made benchmark vignette reproducible (#13)
