+++ data-raw/poset_benchmark.R
@@ -1,0 +1,168 @@
+library(igraph)
+library(netrankr)
+library(progress)
+library(tidyverse)
+library(igraph6)
+
+# load all benchmark posets in digraph6 format
+posets <- readLines("data-raw/posets.txt") 
+
+res_all <- data.frame(id = 1:length(posets),
+                  n = NA,
+                  density = NA,
+                  runtime_exact = 0,
+                  runtime_mcmc = 0,
+                  err_approx_exp_lpom  = 0,
+                  err_approx_exp_glpom = 0,
+                  err_approx_exp_loof1 = 0,
+                  err_approx_exp_loof2 = 0,
+                  err_approx_exp_mcmc  = 0,
+                  err_approx_rel_it1 = 0,
+                  err_approx_rel_it5 = 0,
+                  err_approx_rel_it10 = 0,
+                  err_approx_rel_it15 = 0,
+                  err_approx_rel_mcmc = 0,
+                  rtime_approx_exp_lpom  = 0,
+                  rtime_approx_exp_glpom = 0,
+                  rtime_approx_exp_loof1 = 0,
+                  rtime_approx_exp_loof2 = 0,
+                  rtime_approx_rel_it1 = 0,
+                  rtime_approx_rel_it5 = 0,
+                  rtime_approx_rel_it10 = 0,
+                  rtime_approx_rel_it15 = 0,
+                  dgraph6 = posets)
+
+total <- length(posets)
+
+pb <- progress_bar$new(
+  format = "[:bar] (:current/:total) eta: :eta",
+  total = total, clear = FALSE, width= 100)
+
+for(k in 1:total){
+  pb$tick()
+  po <- posets[k]
+  A <- rgraph6::adjacency_from_digraph6(po)[[1]]
+  res_all$n[k] <- nrow(A)
+  res_all$density[k] <- comparable_pairs(A)
+  if(res_all$density[k]==1){
+    next()
+  }
+  #exact
+  res_all$runtime_exact[k] <- system.time(res_exact <- exact_rank_prob(A))[3]
+  #mcmc
+  res_all$runtime_mcmc[k] <- system.time(res_mcmc <- netrankr::mcmc_rank_prob(A,nrow(A)^5))[3]
+  #approx expected
+  res_all$rtime_approx_exp_lpom[k] <- system.time(lpom  <- approx_rank_expected(A, method="lpom"))[3]
+  res_all$rtime_approx_exp_glpom[k] <- system.time(glpom <- approx_rank_expected(A, method="glpom"))[3]
+  res_all$rtime_approx_exp_loof1[k] <- system.time(loof1 <- approx_rank_expected(A, method="loof1"))[3]
+  res_all$rtime_approx_exp_loof2[k] <- system.time(loof2 <- approx_rank_expected(A, method="loof2"))[3]
+  #approx relative
+  res_all$rtime_approx_rel_it1[k] <- system.time(rel_it1 <- approx_rank_relative(A, iterative = FALSE))[3]
+  res_all$rtime_approx_rel_it5[k] <- system.time(rel_it5 <- approx_rank_relative(A, iterative = TRUE,num.iter = 5))[3]
+  res_all$rtime_approx_rel_it10[k] <- system.time(rel_it10 <-approx_rank_relative(A, iterative = TRUE,num.iter = 10) )[3]
+  res_all$rtime_approx_rel_it15[k] <- system.time(rel_it15 <- approx_rank_relative(A, iterative = TRUE,num.iter = 15))[3]
+  
[…]
