+++ data-raw/posets.txt
@@ -1,0 +1,3000 @@
+&IF{^p~B{FoN?[?O@??
+&IC_@_G??A?????????
+&I????CA???B???????
+&IA?G?@AG????O?????
+&I^{~p~@{FoN?[?o@??
+&IF{~p~A{?_J?C?O???
+&I^{Np~Ak?_J?G?O???
+&I^{\pZ?C@oB?K?o@??
+&I^{~p~B{FoN?[?o@??
+&IN{~o~B{FoL?S?O@??
+&IJw^o}B{F_M?W?????
+&I?O???????????????
+&IN{~o~?[??N?[?o???
+&IF{~p~B{EoJ?K?o@??
+&I^{^p~B{FoN?[?O@??
+&IV{^oHB{FoN?C?O@??
+&I^{^o~B{FoN?O??@??
+&I^{^p~@{FoN?C?o???
+&I??@ODAsBO??S?O???
+&I^{~p~B{FoN?K?O@??
+&I??????OA?????????
+&I^{No~B{FoF?[?o???
+&I^{~p~B{@oJ?K?o@??
+&I@O_??AG????O?????
+&IB{^o~@{FoL?S??@??
+&I^{~p~@[FOD?S?O???
+&I^{~p~B{FoN?[?o@??
+&IN{^o^B{FoN?[?o???
+&IGg?oA??@?????????
+&I^{^p^@kFoJ???O@??
+&I^{~p~B{FoN?[?o@??
+&IN{ZoVA{?OD?[?O@??
+&I?S??@??A?????????
+&I?W@_I@?A_??G?_???
+&IV[\oP??Do??W?_???
+&I^{~o^B{FoN?[?O@??
+&IF[\oDBKCo??G?????
+&I^{~o~B{FoN?K?o???
+&IN{~p~B{FoN?[?o@??
+&IF{^p~@kAoJ?K?o@??
+&IA????????????????
+&I?G????C??????????
+&I@C??C?G??????????
+&I?OT?b@O?O?????@??
+&I\{voE@kCO??C?????
+&IV{^o~?[FoN?[?o???
+&I@[GofA[?OE?G?_???
+&I^{Np~B{FoF?K?o@??
+&I^{~p~B{FoN?[?o@??
+&IN{~p~B{@oB?[??@??
+&I?kNoRA{BoB?C?o@??
+&I?????????????????
+&I??P?c?O????C?????
+&IJ{~oN?{DoB?[?o@??
+&I^{rpN?[BoN?K?o@??
+&I\{roE??C_A?[?_???
+&I^{~o~B{FoN?[?o@??
+&I?????????????????
+&IC?A_???CO??G?????
+&I@Gc_QB{E_A?G?O???
+&IAS??D??@O@?G?O???
+&IN{~p~B{FoN?K?o@??
+&IN{~p~B{FoN?[?o@??
+&IMw??_?G????W?_???
+&I?????????????O???
+&IJwO_C?G??A???????
+&IHC_???gFo@?G?????
+&IZkmOG@_AO????_???
+&IB[DoeA[@_@???_???
+&IZ{Jo~B{BoN?[?o???
+&IR{No^BkEoJ?K?O@??
+&IPKCoR?sBo@?O?????
+&I^{~p~A{BoN?[?O???
+&IE{AO???Ao????O@??
+&I^{~p~B{FoN?[?o@??
+&I^{~p~B{@oF?[?o@??
+&I^{^o^B{DoF?[??@??
+&I???????@???C?????
+&ICwE???_E_G???????
+&I^{no~B{FoF?[?o@??
+&I^{no^B{BoN?[?o@??
[…]
