### 2022-09-20 schochastics/netrankr#13 @schochastics
This is a very good point that I missed. I added replication material in the `data-raw` folder. This is not the material I used back when I added the vignette (I was oblivious about reproducible code then) but from a more recent project. The results dont change, but I did change the figures in the vignette to reflect the different benchmark dataset.
