### 2022-09-20 bdd1f7f
added Contributing guidelines (#16)
