### 2022-09-20 schochastics/netrankr#16 @schochastics
Added guidelines under .github
