### 2022-09-20 schochastics/netrankr#12 @schochastics
This is a very good point and actually has been misinterpreted by some. I will change this in the next CRAN submission
