### 2022-09-20 c8798a5
labeld toy example as such (#14)
