### 2022-09-20 schochastics/netrankr#12 @corybrunson
Aha, good to know, and thank you. I'll leave the issue for you to close upon this change. : )
