### 2022-09-20 schochastics/netrankr#17 @schochastics
I just learned about the `goodpractice` paper a week ago or so. I'll incorporate the suggestions for the next release
