### 2022-09-20 schochastics/netrankr#13 @corybrunson
This is great, thank you! I only further request that the benchmarking vignette tell users where to find this script so that they can run and adapt it as needed.
