### 2022-09-20 #4563 @schochastics
I addressed issues 13 and 16. I think 16 is already resolved and now just waiting for @corybrunson to confirm the reproducibility of benchmark results.

The only thing that was changed in the paper is to explicitly state that the graph is a toy example specifically crafted for the purpose of illustration.
