### 2022-09-21 4f8e47f
fixed some goodpractice suggestions #17
