+++ paper.md
@@ -36,31 +36,30 @@ of centrality based on partial and probabilistic rankings in networks.
 
 # Statement of need
 
-General purpose packages for network analysis such as `igraph` [@cn-ispcnr-06] or
- `sna` [@b-snas-08], implement the most frequently used centrality indices.
+General purpose packages for network analysis such as `igraph` [@cn-ispcnr-06] and
+ `sna` [@b-snas-08] implement methods for the most frequently used centrality indices.
 
 However, there also exist dedicated centrality packages, such as 
 `centiserve` [@jsaayga-ccrwarpca-15], `CINNA` [@amj-crpdcinna-19], and `influenceR` [@sa-istqsinn-15].
 The biggest in terms of implemented indices is `centiserve` which includes $33$ indices.
 The primary purpose of `CINNA` is to facilitate the choice of indices by visual 
-and statistical tools. `influenceR` is a comparably small package which implements 
+and statistical tools. `influenceR` is a relatively tiny package which implements 
 only a few specialized measures.
 
 The `netrankr` package also offers 
 the opportunity to apply more than $30$ indices, but the main focus lies on alternative assessment methods.
-Recent results suggest, that there exist a 
-variety of partial rankings in networks which are preserved by many index based centrality rankings [@sb-sninc-15; @sb-rcsn-16; @svb-ccicurg-17]. 
+Recent results suggest that a variety of partial rankings exists in networks which are preserved by many index based centrality rankings [@sb-sninc-15; @sb-rcsn-16; @svb-ccicurg-17]. 
 These partial rankings can be leveraged to assess centrality on a more general level, without necessarily resorting to indices. 
 Some of these methods and key functionalities of the package are listed below.
 
 * Compute partial rankings based on neighborhood-inclusion and other dominance relations in networks [@sb-sninc-15; @b-np-16]
 * Analyze the partial rankings using rank intervals [@pt-miposlemrp-04] 
 * Calculate rank probabilities to assess the likelihood of certain ranks (e.g. How likely is a node the most central one?) [@s-ciprrpn-18]
-* Calculate mutual rank probabilities (e.g. how likely is a node more central than another one?)[@s-ciprrpn-18]
-* Compute expected ranks of nodes (How central do we expect a node to be?)[@s-ciprrpn-18]
+* Calculate mutual rank probabilities (e.g. How likely is a node more central than another one?)[@s-ciprrpn-18]
+* Compute expected ranks of nodes (e.g. How central do we expect a node to be?)[@s-ciprrpn-18]
 
-Note that most of the tools can also be applied in other empirical settings where partial orders 
-arise and need to be analyzed.
+Note that the most of the tools can also be applied in other empirical settings where partial orders 
+arise.
 
 # Background
 
@@ -68,15 +67,15 @@ In an undirected graph $G=(V,E)$ with vertex set $V$ (with cardinality $n = \lve
 of a node $u \in V$ is defined as
 $$N(u)=\lbrace w : \lbrace u,w \rbrace \in E \rbrace$$
 and its closed neighborhood as $N[v]=N(v) \cup \lbrace v \rbrace$. If the 
-neighborhood of a node $u$ is a subset of the closed neighborhood of a node 
-$v$, $N(u)\subseteq N[v]$, we speak of neighborhood inclusion. This concept 
-defines a binary relation among nodes in a network and we say that $u$ is 
[…]
