### 2022-09-23 #4563 @jbytecode
@schochastics - I have just sent a pull request which contains some text correction in paper.md

Please review the changes and apply the pull request if you are agreed with them.

Please consider these task:

- [ ] the figure `figures/dbces.pdf` is not cited/referenced in text and it is not clear how one can generate it and/or which code part is responsible to generate it and/or what does it show. Please set a caption for this figure.
- [ ] Since generating `figures/rk_intervals.pdf` using a `plot()` call, please set a label (caption) for this figure.
- [ ] Please ping me when you have done with them.

Thank you in advance.
