### 2022-09-23 a95ab6e
fixed suggestions mentioned in review https://github.com/openjournals/joss-reviews/issues/4563#issuecomment-1256423623
