### 2022-09-23 #4563 @schochastics
@jbytecode :
- [x] paper checked again, no changes
- [x] tagged version `v1.2.0` [Link](https://github.com/schochastics/netrankr/releases/tag/v1.2.0)
- [x] DOI: https://doi.org/10.5281/zenodo.7109040
- [x] Zenodo Archive: https://zenodo.org/record/7109041
