### 2022-09-25 4027428
removed points from plot since it resulted in too big file for CRAN
