+++ DESCRIPTION
@@ -26,7 +26,7 @@ Imports: igraph (>= 1.0.1),
     Matrix
 LinkingTo: Rcpp,RcppArmadillo
 SystemRequirements: C++11
-RoxygenNote: 7.2.0
+RoxygenNote: 7.2.1
 Roxygen: list(markdown = TRUE)
 Suggests: knitr, rmarkdown, magrittr,
     testthat,
